[labs/virtualizer] Migrate self-scroller mode to SelfScrollSource and complete ScrollSource refactor (Step 4)
All three scroll modes now flow through a `ScrollSource`:
- `'managed'` → `ManagedScrollSource`
- `true` → `SelfScrollSource`
- `false` → `AncestorScrollSource`

New `BaseDomScrollSource` shared by Self and Ancestor modes encapsulates
the ScrollerController lifecycle, scroll/resize listeners, viewport
math, scroll-error correction, and smooth-scroll orchestration.
Subclasses override three small hooks for the differences.

`ScrollSource` interface gains `checkScrollIntoViewTarget()` so smooth-
scroll retargeting can live in the source. Shared DOM helpers move into
`scroll-sources/_dom-utils.ts`.

Virtualizer is significantly slimmer: removed `_scrollerController`,
`_clippingAncestors`, `_scrollEventListeners`, `_hostElementRO`,
`_windowResizeCallback`, `_scrollIntoViewTarget`,
`_updateScrollIntoViewCoordinates`, `_scrollerWritingMode`,
`_observeAndListen`, `_handleScrollEvent`, `handleEvent`,
`_viewportSizeChanged`, and the inline shadow-DOM-aware helpers.
`_updateView`, `_correctScrollError`, `_scrollElementIntoView`, and
`_checkScrollIntoViewTarget` are now thin delegations to the source.
`_isScroller` is kept for CSS-only decisions.

Net diff: ~890 lines removed from existing files plus the new sources.
All 196 existing tests pass.
